#110 length data type correction use long instead of int

+++ src/main/java/de/cismet/cismap/custom/attributerule/EzgMvWaRuleSet.java
@@ -90,12 +90,12 @@ public class EzgMvWaRuleSet extends WatergisDefaultRuleSet {
 
     @Override
     public Object getAdditionalFieldValue(final java.lang.String propertyName, final FeatureServiceFeature feature) {
-        Integer value = null;
+        Long value = null;
 
         final Geometry geom = ((Geometry)feature.getProperty("geom"));
 
         if (geom != null) {
-            value = (int)Math.round(geom.getArea());
+            value = Math.round(geom.getArea());
         }
 
         return value;
@@ -103,7 +103,7 @@ public class EzgMvWaRuleSet extends WatergisDefaultRuleSet {
 
     @Override
     public Class getAdditionalFieldClass(final int index) {
-        return Integer.class;
+        return Long.class;
     }
 
     @Override
